+++ src/hash_index.rs
@@ -83,6 +83,25 @@ where
     barrier: &'b Barrier,
 }
 
+/// Specifies possible actions of [`HashIndex::modify`] and [`HashIndex::modify_async`] on the
+/// entry.
+pub enum ModifyAction<V> {
+    /// Do nothing.
+    ///
+    /// [`Option::None`] is coerced to [`ModifyAction::Keep`].
+    Keep,
+
+    /// Do remove the entry.
+    ///
+    /// `Some(None)` is coerced to [`ModifyAction::Remove`].
+    Remove,
+
+    /// Do update the entry.
+    ///
+    /// `Some(Some(V))` is coerced to [`ModifyAction::Update`].
+    Update(V),
+}
+
 impl<K, V, H> HashIndex<K, V, H>
 where
     K: 'static + Clone + Eq + Hash,
@@ -247,18 +266,26 @@ where
         }
     }
 
-    /// Replaces the existing value corresponding to the key if the supplied closure constructs a
-    /// new value.
+    /// Modifies the existing entry associated with the specified key.
     ///
-    /// `conditional_constructor` reads the current state of the value associated with the key, and
-    /// then returns `Some(V)` if the current value satisfies any conditions defined in it,
-    /// otherwise returns `None`.
+    /// The return value of the supplied closure denotes the type of an action this method should
+    /// take. The closure may return [`ModifyAction`] or `Option<new_version>` where the type of
+    /// `new_version` is `Option<V>`.
     ///
-    /// Returns `true` if the value was successfully replaced.
+    /// To be specific, this method takes either one of the following actions if the key exists.
+    /// * Do nothing if [`ModifyAction::Keep`] or `None` is returned from the supplied closure.
+    /// * Remove the entry if [`ModifyAction::Remove`] or `Some(None)` is returned from the
+    /// supplied closure.
+    /// * A new version of the entry is created with the current version removed if the supplied
+    /// closure returns [`ModifyAction::Update`] or `Some(Some(new_value))`.
+    ///
+    /// Returns `true` if the entry was either removed or updated; in other words, `false` is
+    /// returned if the key does not exist or the entry was not removed and updated.
     ///
     /// # Examples
     ///
     /// ```
+    /// use scc::hash_index::ModifyAction;
     /// use scc::HashIndex;
     ///
     /// let hashindex: HashIndex<u64, u32> = HashIndex::default();
@@ -266,18 +293,20 @@ where
     /// assert!(unsafe { hashindex.update(&1, |_, _| true).is_none() });
     /// assert!(hashindex.insert(1, 0).is_ok());
     ///
-    /// assert!(!hashindex.replace_if(&1, |_, v| { if *v == 1 { Some(2) } else { None } }));
-    /// assert_eq!(hashindex.read(&1, |_, v| *v).unwrap(), 0);
+    /// assert!(!hashindex.modify(&0, |_, v| Some(Some(1))));
+    /// assert!(hashindex.modify(&1, |_, v| ModifyAction::Update(1)));
+    /// assert_eq!(hashindex.read(&1, |_, v| *v).unwrap(), 1);
     ///
-    /// assert!(hashindex.replace_if(&1, |_, v| { if *v == 0 { Some(2) } else { None } }));
-    /// assert_eq!(hashindex.read(&1, |_, v| *v).unwrap(), 2);
+    /// assert!(hashindex.modify(&1, |_, v| ModifyAction::Remove));
+    /// assert!(hashindex.read(&1, |_, v| *v).is_none());
     /// ```
     #[inline]
-    pub fn replace_if<Q, F>(&self, key: &Q, conditional_constructor: F) -> bool
+    pub fn modify<Q, F, R>(&self, key: &Q, updater: F) -> bool
     where
         K: Borrow<Q>,
         Q: Eq + Hash + ?Sized,
-        F: FnOnce(&K, &V) -> Option<V>,
+        F: FnOnce(&K, &V) -> R,
+        R: Into<ModifyAction<V>>,
     {
         let barrier = Barrier::new();
         let hash = self.hash(key);
@@ -287,49 +316,68 @@ where
         };
         if entry_ptr.is_valid() {
             let (k, v) = entry_ptr.get(data_block_mut);
-            if let Some(new_v) = conditional_constructor(k, v) {
-                let new_k = k.clone();
-                locker.insert_with(
-                    data_block_mut,
-                    BucketArray::<K, V, false>::partial_hash(hash),
-                    || (new_k, new_v),
-                    &barrier,
-                );
+            let modify_action: ModifyAction<V> = updater(k, v).into();
+            let result = match modify_action {
+                ModifyAction::Keep => false,
+                ModifyAction::Remove => true,
+                ModifyAction::Update(new_v) => {
+                    // A new version of the entry will be created.
+                    let new_k = k.clone();
+                    locker.insert_with(
+                        data_block_mut,
+                        BucketArray::<K, V, false>::partial_hash(hash),
+                        || (new_k, new_v),
+                        &barrier,
+                    );
+                    true
+                }
+            };
+            if result {
+                // The entry was modified, and therefore the old version should be logically
+                // removed from the `HashIndex`.
                 locker.erase(data_block_mut, &mut entry_ptr);
-                return true;
             }
+            return result;
         }
         false
     }
 
-    /// Replaces the existing value corresponding to the key if the supplied closure constructs a
-    /// new value.
+    /// Modifies the existing entry associated with the specified key.
     ///
-    /// `conditional_constructor` reads the current state of the value associated with the key, and
-    /// then returns `Some(V)` if the current value satisfies any conditions defined in it,
-    /// otherwise returns `None`.
+    /// The return value of the supplied closure denotes the type of an action this method should
+    /// take. The closure may return [`ModifyAction`] or `Option<new_version>` where the type of
+    /// `new_version` is `Option<V>`.
     ///
-    /// Returns `true` if the value was successfully replaced. It is an asynchronous method returning an
-    /// `impl Future` for the caller to await.
+    /// To be specific, this method takes either one of the following actions if the key exists.
+    /// * Do nothing if [`ModifyAction::Keep`] or `None` is returned from the supplied closure.
+    /// * Remove the entry if [`ModifyAction::Remove`] or `Some(None)` is returned from the
+    /// supplied closure.
+    /// * A new version of the entry is created with the current version removed if the supplied
+    /// closure returns [`ModifyAction::Update`] or `Some(Some(new_value))`.
+    ///
+    /// Returns `true` if the entry was either removed or updated; in other words, `false` is
+    /// returned if the key does not exist or the entry was not removed and updated. It is an
+    /// asynchronous method returning an `impl Future` for the caller to await.
     ///
     /// # Examples
     ///
     /// ```
+    /// use scc::hash_index::ModifyAction;
     /// use scc::HashIndex;
     ///
     /// let hashindex: HashIndex<u64, u32> = HashIndex::default();
     ///
     /// assert!(hashindex.insert(1, 0).is_ok());
-    /// let future_replace = hashindex.replace_if_async(&1, |_, v| {
-    ///     if *v == 0 { Some(2) } else { None }
-    /// });
+    /// let future_update = hashindex.modify_async(&1, |_, v| Some(Some(2)));
+    /// let future_remove = hashindex.modify_async(&1, |_, v| ModifyAction::Remove);
     /// ```
     #[inline]
-    pub async fn replace_if_async<Q, F>(&self, key: &Q, conditional_constructor: F) -> bool
+    pub async fn modify_async<Q, F, R>(&self, key: &Q, updater: F) -> bool
     where
         K: Borrow<Q>,
         Q: Eq + Hash + ?Sized,
-        F: FnOnce(&K, &V) -> Option<V>,
+        F: FnOnce(&K, &V) -> R,
+        R: Into<ModifyAction<V>>,
     {
         let hash = self.hash(key);
         loop {
@@ -342,17 +390,28 @@ where
                 {
                     if entry_ptr.is_valid() {
                         let (k, v) = entry_ptr.get(data_block_mut);
-                        if let Some(new_v) = conditional_constructor(k, v) {
-                            let new_k = k.clone();
-                            locker.insert_with(
-                                data_block_mut,
-                                BucketArray::<K, V, false>::partial_hash(hash),
-                                || (new_k, new_v),
-                                &barrier,
-                            );
+                        let modify_action: ModifyAction<V> = updater(k, v).into();
+                        let result = match modify_action {
+                            ModifyAction::Keep => false,
+                            ModifyAction::Remove => true,
+                            ModifyAction::Update(new_v) => {
+                                // A new version of the entry will be created.
+                                let new_k = k.clone();
+                                locker.insert_with(
+                                    data_block_mut,
+                                    BucketArray::<K, V, false>::partial_hash(hash),
+                                    || (new_k, new_v),
+                                    &barrier,
+                                );
+                                true
+                            }
+                        };
+                        if result {
+                            // The entry was modified, and therefore the old version should be
+                            // logically removed from the `HashIndex`.
                             locker.erase(data_block_mut, &mut entry_ptr);
-                            return true;
                         }
+                        return result;
                     }
                     return false;
                 };
@@ -1265,4 +1324,26 @@ where
     V: 'static + Clone + UnwindSafe,
     H: BuildHasher + UnwindSafe,
 {
+}
+
+impl<V> Debug for ModifyAction<V> {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Keep => write!(f, "Nothing"),
+            Self::Remove => write!(f, "Remove"),
+            Self::Update(_) => f.debug_tuple("Update").finish(),
+        }
+    }
+}
+
+impl<V> From<Option<Option<V>>> for ModifyAction<V> {
+    #[inline]
+    fn from(value: Option<Option<V>>) -> Self {
+        match value {
+            Some(Some(value)) => Self::Update(value),
+            Some(None) => Self::Remove,
+            None => Self::Keep,
+        }
+    }
 }