+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 1.6.3
 
-* API update: add `HashIndex::{replace_if, replace_if_async}`.
+* API update [#96](https://github.com/wvwwvwwv/scalable-concurrent-containers/issues/96): add `HashIndex::{modify, modify_async}`, special thanks to [novacrazy](https://github.com/novacrazy).
 
 1.6.2